requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.3
openai-whisper==20231117
packaging==23.2
pycparser==2.21
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.12.0
sounddevice==0.4.6
sympy==1.12
tiktoken==0.5.2
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
typing-extensions==4.9.0
urllib3==2.2.0
